% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_docs.R
\name{download_all_species_pdfs}
\alias{download_all_species_pdfs}
\title{Download PDFs of documents linked on NMFS species pages}
\usage{
download_all_species_pdfs(subd)
}
\arguments{
\item{subd}{The directory (subdirectory) to which the PDFs are downloaded}
}
\description{
Download PDFs of documents linked on NMFS species pages
}
\note{
This function will only get documents linked from pages linked to
  NMFS's Protected Resources ESA-listed species page,
  \url{http://www.nmfs.noaa.gov/pr/species/esa/listed.htm}. In general this
  means recovery plans and many \emph{Federal Register} documents will not
  be gathered.
}
\examples{
\dontrun{
  download_all_species_pdfs()
}
}

