% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status_reviews.R
\name{download_status_reviews}
\alias{download_status_reviews}
\title{Download all PDF status review documents from NMFS}
\usage{
download_status_reviews(subd = "")
}
\arguments{
\item{subd}{The directory (subdirectory) to which the PDFs are downloaded}
}
\description{
Download all PDF status review documents from NMFS
}
\details{
Uses \link{get_status_review_pdf_urls} to fetch the vector of PDF
URLs for status reviews maintained by the National Marine Fisheries
Service (NMFS). Filenames are the \link{basename} of the URL with spaces
replaced by "_". Uses \link[pdfdown]{pdfdown}, which returns a data.frame of
results, to do the scraping.
}
\note{
Two points to note: \enumerate{
  \item{
    \link[pdfdown]{download_pdf} checks for duplicate files by name and
      skips the download by default if the file exists;
  }
  \item{
      By extension, \link[pdfdown]{download_pdf} does not check for duplicate
      files based on, e.g., MD5 hashes. Yet.
  }
}
}
\examples{
\dontrun{
  dl_res <- download_status_reviews("~/Downloads/NMFS_rec")
}
}

