% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_listings_page.R
\name{get_listings_page}
\alias{get_listings_page}
\title{Read NMFS's ESA listings page}
\usage{
get_listings_page()
}
\description{
Read NMFS's ESA listings page
}
\examples{
\dontrun{
   pg <- get_listings_page()
}
}

