% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery_plans.R
\name{get_plan_urls_direct}
\alias{get_plan_urls_direct}
\title{Return PDF urls linked directly to NMFS's main recovery page}
\usage{
get_plan_urls_direct(page, domain = "http://www.nmfs.noaa.gov")
}
\description{
Return PDF urls linked directly to NMFS's main recovery page
}

