% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery_plans.R
\name{get_plan_urls_indirect}
\alias{get_plan_urls_indirect}
\title{Return PDF urls linked indirectly to NMFS's main recovery page}
\usage{
get_plan_urls_indirect(page)
}
\description{
Return PDF urls linked indirectly to NMFS's main recovery page
}

