% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_policies.R
\name{get_policies_page}
\alias{get_policies_page}
\title{Return NMFS's main policies page as an xml2 object}
\usage{
get_policies_page()
}
\description{
Return NMFS's main policies page as an xml2 object
}

