% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_policies.R
\name{get_policies_pdf_urls}
\alias{get_policies_pdf_urls}
\title{Return PDF urls linked directly to NMFS's main policies page}
\usage{
get_policies_pdf_urls()
}
\description{
Return PDF urls linked directly to NMFS's main policies page
}

