% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_policies.R
\name{get_policies_tables}
\alias{get_policies_tables}
\title{Get tables of NMFS ESA policies}
\usage{
get_policies_tables()
}
\value{
A list of two data.frames, \describe{
  \item{interagency}{Table of interagency cooperation ESA policies}
  \item{other_docs}{Table of other ESA policies}
}
}
\description{
Get tables of NMFS ESA policies
}
\details{
NMFS provides tables with document titles, dates, and citations
for their ESA policy documents. This function may be particularly useful in
a script that is run as a cron job to check for updates.
}
\note{
This function may be fragile: it is 'tuned' to the table layout of
the ESA policies page as of 26 Nov 2016. Check the policies page,
\href{http://www.nmfs.noaa.gov/pr/laws/esa/policies.htm}{here}, for a layout
other than one two-column table at top, followed by >1 three-column tables
below, if this function fails.
}
\examples{
\dontrun{
   policies_tab <- get_policies_tables()
}
}

