% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery_plans.R
\name{get_recovery_plan_pdf_urls}
\alias{get_recovery_plan_pdf_urls}
\title{Return PDF urls for ESA recovery plans from NMFS}
\usage{
get_recovery_plan_pdf_urls()
}
\value{
A vector of URLs for recovery plan PDFs
}
\description{
Return PDF urls for ESA recovery plans from NMFS
}
\details{
The National Marine Fisheries Service (NMFS) provides recovery plan
materials (the plans, addendums, etc.) for Endangered Species Act-listed
(ESA) species under its jurisdiction at
\url{http://www.nmfs.noaa.gov/pr/recovery/plans.htm}. Except that not all
plans or plan materials are located at that site; plans for different
Chinook salmon populations are located elsewhere. This function returns a
vector of all PDF URLs on the main recovery plan page as well as PDFs on
pages linked from the main page. After getting this list, all of the PDFs
can be downloaded using a custom scraper, with \link[pdfdown]{pdfdown}, or
using \link{download_recovery_plans}.
}

