% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery_plans.R
\name{get_recovery_plan_urls_indirect}
\alias{get_recovery_plan_urls_indirect}
\title{Return PDF urls linked indirectly to NMFS's main recovery page}
\usage{
get_recovery_plan_urls_indirect(page)
}
\arguments{
\item{page}{An \link[xml2]{read_html} page}
}
\description{
Return PDF urls linked indirectly to NMFS's main recovery page
}
\details{
While most species' recovery plans are linked directly to the main
NMFS recovery page, some are hosted at regional offices or science centers.
This function collects these non-central PDF URLs.
}

