% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery_plans.R
\name{get_recovery_plans}
\alias{get_recovery_plans}
\title{Download all PDF recovery documents from NMFS}
\usage{
get_recovery_plans(subd = "")
}
\arguments{
\item{subd}{The directory (subdirectory) to which the PDFs are downloaded}
}
\description{
Download all PDF recovery documents from NMFS
}
\details{
Uses \link{get_plan_pdf_urls} to fetch the vector of PDF URLs for
recovery plans maintained by the National Marine Fisheries Service (NMFS).
Filenames are the \link{basename} of the URL with spaces replaced by "_".
Uses \link[pdfdown]{pdfdown}, which returns a data.frame of results, to
do the scraping. Note that \link[pdfdown]{download_pdf} will not
}
\examples{
\dontrun{
dl_res <- get_recovery_plans("~/Downloads/NMFS_rec")
}
}

