% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recovery_plans.R
\name{get_recovery_plans_page}
\alias{get_recovery_plans_page}
\title{Return NMFS's main recovery plan page as an xml2 object}
\usage{
get_recovery_plans_page()
}
\description{
Return NMFS's main recovery plan page as an xml2 object
}

