% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_docs.R
\name{get_species_pdf_urls}
\alias{get_species_pdf_urls}
\title{Get all PDF URLs from query URL}
\usage{
get_species_pdf_urls(url)
}
\arguments{
\item{url}{The URL to be queried for PDF links}
}
\description{
Function description; defaults to title if left blank
}
\examples{
\dontrun{
   url <- "http://www.nmfs.noaa.gov/pr/species/turtles/green.html"
   chemyd_pdfs <- get_species_pdf_urls(url)
}
}

