% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_species_docs.R
\name{get_species_pdfs}
\alias{get_species_pdfs}
\title{Download all PDF recovery documents from NMFS}
\usage{
get_species_pdfs(subd = "")
}
\arguments{
\item{subd}{The directory (subdirectory) to which the PDFs are downloaded}
}
\description{
Download all PDF recovery documents from NMFS
}
\details{
Uses \link{get_plan_pdf_urls} to fetch the vector of PDF URLs for
recovery plans maintained by the National Marine Fisheries Service (NMFS).
Filenames are the \link{basename} of the URL with spaces replaced by "_".
Uses \link[pdfdown]{pdfdown}, which returns a data.frame of results, to
do the scraping. Note that \link[pdfdown]{download_pdf} will not
}
\examples{
\dontrun{
dl_res <- get_species_pdfs("~/Downloads/NMFS_rec")
}
}

