% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status_reviews.R
\name{get_status_review_page}
\alias{get_status_review_page}
\title{Return NMFS's main status review page as an xml2 object}
\usage{
get_status_review_page()
}
\description{
Return NMFS's main status review page as an xml2 object
}

