% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_status_reviews.R
\name{get_status_review_pdf_urls}
\alias{get_status_review_pdf_urls}
\title{Return PDF urls linked directly to NMFS's main status reviews page}
\usage{
get_status_review_pdf_urls()
}
\description{
Return PDF urls linked directly to NMFS's main status reviews page
}

