% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMFS_glossary.R
\docType{data}
\name{NMFS_glossary}
\alias{NMFS_glossary}
\title{Glossary of terms from the National Marine Fisheries Service (NMFS)}
\format{A data frame with 877 rows and 2 variables:
\describe{
  \item{term}{the term to be defined}
  \item{definition}{the definition of the term}
}}
\source{
\url{http://www.st.nmfs.noaa.gov/st4/documents/FishGlossary.pdf}
}
\usage{
NMFS_glossary
}
\description{
Definitions for terms commonly used by the National Marine Fisheries Service.
}
\keyword{datasets}

