% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmfsscraper.R
\docType{package}
\name{nmfsscraper}
\alias{nmfsscraper}
\alias{nmfsscraper-package}
\title{Scrape ESA documents from the NMFS website}
\description{
Scrape ESA documents from the NMFS website
}
\details{
The National Marine Fisheries Service (NMFS), along with the U.S.
Fish and Wildlife Service (FWS), is responsible for implementing the
Endangered Species Act (ESA). There is a huge amount of information about
ESA-listed species and about how the ESA is implemented in tens of thousands
of PDFs that the Services host, but the Services don't make those documents
available from a single, central download location. This package is a (set
of) scrapers to search for PDFs on the NMFS website,
\url{http://www.nmfs.noaa.gov}, and download them locally to facilitate
analysis of the embedded information, e.g., using Natural Language Processing
(NLP).
}

